3619: Rename AttributeValueWithDoubleQuotationMarksIssueGenerator files
