+++ common/src/Model/PropertyValueWithDoubleQuotationMarksIssueGenerator.h
